refactor: move cicd parser under yaml parser
