+++ pkg/parser/yaml/cicd/expression_parser_test.go
