+++ pkg/parser/yaml/cicd/parser.go
