+++ pkg/parser/yaml/cicd/parser_test.go
