+++ pkg/parser/yaml/cicd/shell_parser.go
