+++ pkg/parser/yaml/cicd/shell_parser_test.go
